stream interface improvement
