+++ p2p/src/authenticated/actors/peer/mod.rs
@@ -1,7 +1,6 @@
 //! Peer
 
 use crate::authenticated::metrics;
-use commonware_stream::public_key;
 use governor::Quota;
 use prometheus_client::metrics::{counter::Counter, family::Family};
 use prost::DecodeError;
@@ -30,11 +29,11 @@ pub enum Error {
     #[error("peer killed: {0}")]
     PeerKilled(String),
     #[error("send failed: {0}")]
-    SendFailed(public_key::Error),
+    SendFailed(commonware_stream::Error),
     #[error("peer disconnected")]
     PeerDisconnected,
     #[error("receive failed: {0}")]
-    ReceiveFailed(public_key::Error),
+    ReceiveFailed(commonware_stream::Error),
     #[error("decode failed: {0}")]
     DecodeFailed(DecodeError),
     #[error("unexpected failure: {0}")]
