+++ p2p/src/authenticated/actors/spawner/ingress.rs
@@ -1,13 +1,13 @@
 use crate::authenticated::actors::tracker;
 use commonware_cryptography::{PublicKey, Scheme};
 use commonware_runtime::{Clock, Sink, Spawner, Stream};
-use commonware_stream::public_key::Instance;
+use commonware_stream::public_key::Connection;
 use futures::{channel::mpsc, SinkExt};
 
 pub enum Message<E: Spawner + Clock, C: Scheme, Si: Sink, St: Stream> {
     Spawn {
         peer: PublicKey,
-        connection: Instance<C, Si, St>,
+        connection: Connection<C, Si, St>,
         reservation: tracker::Reservation<E>,
     },
 }
@@ -24,7 +24,7 @@ impl<E: Spawner + Clock, C: Scheme, Si: Sink, St: Stream> Mailbox<E, C, Si, St> 
     pub async fn spawn(
         &mut self,
         peer: PublicKey,
-        connection: Instance<C, Si, St>,
+        connection: Connection<C, Si, St>,
         reservation: tracker::Reservation<E>,
     ) {
         self.sender
